修改result类

+++ backend/src/main/java/com/music/common/lang/PageResult.java
@@ -8,20 +8,20 @@ public class PageResult implements Serializable {
     private int code;
     private String msg;
     private Object data;
-    private int total;
+    private long total;
     public static PageResult succ(Object data) {
         return succ(200,"操作成功",data,0);
     }
 
-    public static PageResult succ(Object data, int total) {
+    public static PageResult succ(Object data, long total) {
         return succ(200,"操作成功",data,total);
     }
 
-    public static PageResult succ(String msg,Object data, int total){
+    public static PageResult succ(String msg,Object data, long total){
         return succ(200,msg,data,total);
     }
 
-    public static PageResult succ(int code, String msg, Object data ,int total) {
+    public static PageResult succ(int code, String msg, Object data ,long total) {
         PageResult r = new PageResult();
         r.setCode(code);
         r.setMsg(msg);
@@ -34,15 +34,15 @@ public class PageResult implements Serializable {
         return fail(400,msg,null,0);
     }
 
-    public static PageResult fail(String msg, int total) {
+    public static PageResult fail(String msg, long total) {
         return fail(400,msg,null,total);
     }
 
-    public static PageResult fail(String msg, Object data, int total) {
+    public static PageResult fail(String msg, Object data, long total) {
         return fail(400,msg,data,total);
     }
 
-    public static PageResult fail(int code, String msg, Object data, int total) {
+    public static PageResult fail(int code, String msg, Object data, long total) {
         PageResult r = new PageResult();
         r.setCode(code);
         r.setMsg(msg);
